Refactor TrouterUtils & chat.swift
